# ui/src/api/openapi-spec.json
{
    "openapi": "3.1.0",
    "info": {
        "title": "Geti Tune",
        "description": "Geti Tune allows to fine-tune computer vision models at the edge. It is a lightweight application that covers the complete model AI lifecycle, including advanced features for inference, model monitoring and data collection.",
        "version": "0.1.0"
    },
    "paths": {
...
        "/api/pipelines/{pipeline_id}": {
            "get": {
                "tags": ["Pipelines"],
                "summary": "Get Pipeline",
                "description": "Get info about a given pipeline",
                "operationId": "get_pipeline_api_pipelines__pipeline_id__get",
                "parameters": [
                    {
                        "name": "pipeline_id",
                        "in": "path",
                        "required": true,
                        "schema": { "type": "string", "title": "Pipeline Id" }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Pipeline found",
                        "content": { "application/json": { "schema": { "$ref": "#/components/schemas/Pipeline" } } }
                    },
                    "400": { "description": "Invalid pipeline ID" },
                    "404": { "description": "Pipeline not found" },
                    "422": {
                        "description": "Validation Error",
                        "content": {
                            "application/json": { "schema": { "$ref": "#/components/schemas/HTTPValidationError" } }
                        }
                    }
                }
            },
            "patch": {
                "tags": ["Pipelines"],
                "summary": "Update Pipeline",
                "description": "Reconfigure an existing pipeline",
                "operationId": "update_pipeline_api_pipelines__pipeline_id__patch",
                "parameters": [
                    {
                        "name": "pipeline_id",
                        "in": "path",
                        "required": true,
                        "schema": { "type": "string", "title": "Pipeline Id" }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "type": "object",
                                "additionalProperties": true,
                                "description": "\nPartial pipeline configuration update. May contain any subset of fields including 'name', 'source_id', \n'sink_id', or 'model_id'. Fields not included in the request will remain unchanged.\n",
                                "title": "Pipeline Config"
                            },
                            "examples": {
                                "switch_model": {
                                    "summary": "Switch active model",
                                    "description": "Change the active model for the pipeline",
                                    "value": { "model_id": "c1feaabc-da2b-442e-9b3e-55c11c2c2ff3" }
                                },
                                "reconfigure": {
                                    "summary": "Reconfigure pipeline",
                                    "description": "Change the name, source and sink of the pipeline",
                                    "value": {
                                        "name": "Updated Production Pipeline",
                                        "source_id": "e3cbd8d0-17b8-463e-85a2-4aaed031674e",
                                        "sink_id": "c6787c06-964b-4097-8eca-238b8cf79fc9"
                                    }
                                }
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Pipeline successfully updated",
                        "content": { "application/json": { "schema": { "$ref": "#/components/schemas/Pipeline" } } }
                    },
                    "400": { "description": "Invalid pipeline ID or request body" },
                    "404": { "description": "Pipeline not found" },
                    "409": { "description": "Pipeline cannot be updated" },
                    "422": {
                        "description": "Validation Error",
                        "content": {
                            "application/json": { "schema": { "$ref": "#/components/schemas/HTTPValidationError" } }
                        }
                    }
                }
            },
            "delete": {
                "tags": ["Pipelines"],
                "summary": "Delete Pipeline",
                "description": "Delete a pipeline. Pipelines must be first disabled (status must be idle) before deletion.",
                "operationId": "delete_pipeline_api_pipelines__pipeline_id__delete",
                "parameters": [
                    {
                        "name": "pipeline_id",
                        "in": "path",
                        "required": true,
                        "schema": { "type": "string", "title": "Pipeline Id" }
                    }
                ],
                "responses": {
                    "204": { "description": "Pipeline successfully deleted" },
                    "400": { "description": "Invalid pipeline ID" },
                    "404": { "description": "Pipeline not found" },
                    "409": { "description": "Pipeline is currently in running state and cannot be deleted" },
                    "422": {
                        "description": "Validation Error",
                        "content": {
                            "application/json": { "schema": { "$ref": "#/components/schemas/HTTPValidationError" } }
                        }
                    }
                }
            }
        },
...
        "/api/models": {
            "post": {
                "tags": ["Models"],
                "summary": "Add Model",
                "description": "Upload a new model\n\nNOTE: this endpoint will be replaced by preconfigured model selection",
                "operationId": "add_model_api_models_post",
                "parameters": [
                    {
                        "name": "model_name",
                        "in": "query",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "description": "Name for the model files",
                            "title": "Model Name"
                        },
                        "description": "Name for the model files"
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "multipart/form-data": {
                            "schema": { "$ref": "#/components/schemas/Body_add_model_api_models_post" }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Successful Response",
                        "content": { "application/json": { "schema": { "$ref": "#/components/schemas/Model" } } }
                    },
                    "422": {
                        "description": "Validation Error",
                        "content": {
                            "application/json": { "schema": { "$ref": "#/components/schemas/HTTPValidationError" } }
                        }
                    }
                }
            },
            "get": {
                "tags": ["Models"],
                "summary": "List Models",
                "description": "Get information about available models",
                "operationId": "list_models_api_models_get",
                "responses": {
                    "200": {
                        "description": "List of available models",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": { "$ref": "#/components/schemas/Model" },
                                    "title": "Response 200 List Models Api Models Get"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/models/{model_id}": {
            "get": {
                "tags": ["Models"],
                "summary": "Get Model",
                "description": "Get information about a specific model",
                "operationId": "get_model_api_models__model_id__get",
                "parameters": [
                    {
                        "name": "model_id",
                        "in": "path",
                        "required": true,
                        "schema": { "type": "string", "title": "Model Id" }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Model found",
                        "content": { "application/json": { "schema": { "$ref": "#/components/schemas/Model" } } }
                    },
                    "400": { "description": "Invalid model ID" },
                    "404": { "description": "Model not found" },
                    "422": {
                        "description": "Validation Error",
                        "content": {
                            "application/json": { "schema": { "$ref": "#/components/schemas/HTTPValidationError" } }
                        }
                    }
                }
            },
            "patch": {
                "tags": ["Models"],
                "summary": "Update Model Metadata",
                "description": "Update the metadata of an existing model",
                "operationId": "update_model_metadata_api_models__model_id__patch",
                "parameters": [
                    {
                        "name": "model_id",
                        "in": "path",
                        "required": true,
                        "schema": { "type": "string", "title": "Model Id" }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": { "type": "object", "additionalProperties": true, "title": "Model Metadata" },
                            "examples": {
                                "rename_model": {
                                    "summary": "Rename model",
                                    "description": "Change the name of the model",
                                    "value": { "name": "New Model Name" }
                                }
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Model successfully updated",
                        "content": { "application/json": { "schema": { "$ref": "#/components/schemas/Model" } } }
                    },
                    "400": { "description": "Invalid model ID or request body" },
                    "404": { "description": "Model not found" },
                    "422": {
                        "description": "Validation Error",
                        "content": {
                            "application/json": { "schema": { "$ref": "#/components/schemas/HTTPValidationError" } }
                        }
                    }
                }
            },
            "delete": {
                "tags": ["Models"],
                "summary": "Delete Model",
                "description": "Delete a model",
                "operationId": "delete_model_api_models__model_id__delete",
                "parameters": [
                    {
                        "name": "model_id",
                        "in": "path",
                        "required": true,
                        "schema": { "type": "string", "title": "Model Id" }
                    }
                ],
                "responses": {
                    "204": { "description": "Model configuration successfully deleted" },
                    "400": { "description": "Invalid model ID" },
                    "404": { "description": "Model not found" },
                    "409": { "description": "Model is used by at least one pipeline" },
                    "422": {
                        "description": "Validation Error",
                        "content": {
                            "application/json": { "schema": { "$ref": "#/components/schemas/HTTPValidationError" } }
                        }
                    }
                }
            }
        },
        "/api/models/{model_name}:activate": {
            "post": {
                "tags": ["Models"],
                "summary": "Activate Model",
                "description": "Activate a model\n\nNOTE: this endpoint will be removed; use instead `PATCH /api/pipelines/{pipeline_id}` to change the active model",
                "operationId": "activate_model_api_models__model_name__activate_post",
                "deprecated": true,
                "parameters": [
                    {
                        "name": "model_name",
                        "in": "path",
                        "required": true,
                        "schema": { "type": "string", "title": "Model Name" }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful Response",
                        "content": { "application/json": { "schema": { "$ref": "#/components/schemas/Model" } } }
                    },
                    "422": {
                        "description": "Validation Error",
                        "content": {
                            "application/json": { "schema": { "$ref": "#/components/schemas/HTTPValidationError" } }
                        }
                    }
                }
            }
        },
        "/api/models/{model_name}:deactivate": {
            "post": {
                "tags": ["Models"],
                "summary": "Deactivate Model",
                "description": "Deactivate a model\n\nNOTE: this endpoint will be removed; use instead `PATCH /api/pipelines/{pipeline_id}` to change the active model",
                "operationId": "deactivate_model_api_models__model_name__deactivate_post",
                "deprecated": true,
                "parameters": [
                    {
                        "name": "model_name",
                        "in": "path",
                        "required": true,
                        "schema": { "type": "string", "title": "Model Name" }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful Response",
                        "content": { "application/json": { "schema": { "$ref": "#/components/schemas/Model" } } }
                    },
                    "422": {
                        "description": "Validation Error",
                        "content": {
                            "application/json": { "schema": { "$ref": "#/components/schemas/HTTPValidationError" } }
                        }
                    }
                }
            }
        },
...
    },
    "components": {
        "schemas": {
...
            "Body_add_model_api_models_post": {
                "properties": {
                    "xml_file": { "type": "string", "format": "binary", "title": "Xml File" },
                    "bin_file": { "type": "string", "format": "binary", "title": "Bin File" }
                },
                "type": "object",
                "required": ["xml_file", "bin_file"],
                "title": "Body_add_model_api_models_post"
            },
...
            "HTTPValidationError": {
                "properties": {
                    "detail": {
                        "items": { "$ref": "#/components/schemas/ValidationError" },
                        "type": "array",
                        "title": "Detail"
                    }
                },
                "type": "object",
                "title": "HTTPValidationError"
            },
...
            "Model": {
                "properties": {
                    "id": { "type": "string", "format": "uuid", "title": "Id" },
                    "name": { "type": "string", "title": "Name", "default": "Default Name" },
                    "format": { "$ref": "#/components/schemas/ModelFormat", "default": "openvino_ir" }
                },
                "type": "object",
                "title": "Model",
                "description": "Base model schema that includes common fields for all models.\nThis can be extended by other schemas to include additional fields.",
                "example": {
                    "format": "openvino_ir",
                    "id": "76e07d18-196e-4e33-bf98-ac1d35dca4cb",
                    "name": "YOLO-X for Vehicle Detection"
                }
            },
            "ModelFormat": { "type": "string", "enum": ["openvino_ir", "onnx"], "title": "ModelFormat" },
...
            "Pipeline": {
                "properties": {
                    "id": { "type": "string", "format": "uuid", "title": "Id" },
                    "name": { "type": "string", "title": "Name", "default": "Default Name" },
                    "source_id": {
                        "anyOf": [{ "type": "string", "format": "uuid" }, { "type": "null" }],
                        "title": "Source Id"
                    },
                    "sink_id": {
                        "anyOf": [{ "type": "string", "format": "uuid" }, { "type": "null" }],
                        "title": "Sink Id"
                    },
                    "model_id": {
                        "anyOf": [{ "type": "string", "format": "uuid" }, { "type": "null" }],
                        "title": "Model Id"
                    },
                    "status": { "$ref": "#/components/schemas/PipelineStatus", "default": "idle" }
                },
                "type": "object",
                "title": "Pipeline",
                "example": {
                    "id": "a1b2c3d4-e5f6-7890-abcd-ef1234567890",
                    "model_id": "b0feaabc-da2b-442e-9b3e-55c11c2c2ff2",
                    "name": "Production Pipeline",
                    "sink_id": "b5787c06-964b-4097-8eca-238b8cf79fc8",
                    "source_id": "d2cbd8d0-17b8-463e-85a2-4aaed031674d",
                    "status": "running"
                }
            },
            "PipelineStatus": { "type": "string", "enum": ["idle", "running"], "title": "PipelineStatus" },
...
            "ValidationError": {
                "properties": {
                    "loc": {
                        "items": { "anyOf": [{ "type": "string" }, { "type": "integer" }] },
                        "type": "array",
                        "title": "Location"
                    },
                    "msg": { "type": "string", "title": "Message" },
                    "type": { "type": "string", "title": "Error Type" }
                },
                "type": "object",
                "required": ["loc", "msg", "type"],
                "title": "ValidationError"
            },
...
        }
    }
}